Enable automatic export downloads
